refactor(datavisualisations): move charts into nested directories
