+++ lib/datavisualisations/BarChart/BarChart.vue
