+++ lib/datavisualisations/BarChart/StackedBarChart.vue
