+++ lib/datavisualisations/index.ts
@@ -1,5 +1,5 @@
-export { default as BarChart } from './BarChart.vue'
-export { default as ColumnChart } from './ColumnChart.vue'
-export { default as LineChart } from './LineChart.vue'
-export { default as StackedBarChart } from './StackedBarChart.vue'
-export { default as StackedColumnChart } from './StackedColumnChart.vue'
+export { default as BarChart } from './BarChart/BarChart.vue'
+export { default as ColumnChart } from './ColumnChart/ColumnChart.vue'
+export { default as LineChart } from './LineChart/LineChart.vue'
+export { default as StackedBarChart } from './BarChart/StackedBarChart.vue'
+export { default as StackedColumnChart } from './ColumnChart/StackedColumnChart.vue'
